style(gc): rustfmt and split the four files this change pushed past the size gate
`cargo fmt` output plus the `scripts/check_file_size.sh` splits this change
needs: it took `object/mod.rs` 1998 -> 2208, `object/descriptor_state.rs`
1815 -> 2044, `gc/roots.rs` 1994 -> 2027 and `gc/cycle.rs` 1998 -> 2023, and
three of those four sat within two lines of the 2000-line limit on main.

Each split follows the gate's own recipe (extract a function group into a
sibling, re-export by name) and each is a group that already read as a unit:

* `object/side_table_roots.rs` — the transition-cache and shape-cache root
  scanners and dead-owner prunes, now four full-walk/minor-scoped pairs.
* `object/descriptor_state/young.rs` — the minor-scoped descriptor walk and
  the re-derivation of the relevant set rule 2 checks it against.
* `gc/roots/stack_bottom.rs` — the four `#[cfg]` arms of `get_stack_bottom`.
  The doc comment on the first arm describes a trace-phase mark helper rather
  than `get_stack_bottom`; it was already attached to that item and moves with
  it verbatim rather than being re-pointed at the next item.
* `gc/cycle/registered_root_scan.rs` — the two registered-root scan cursors
  the budgeted root scan resumes through.

`scripts/gc_rekeyed_key_tables.json` follows `scan_transition_cache_slot` to
its new file (the gate reported it as one UNCLASSIFIED site and one STALE
entry, which is the gate working), and the two `#[cfg(test)]` transition-cache
seams are re-exported for `gc::tests::dead_owner_side_tables`.

No behaviour change: every moved item keeps its body, and visibility widens
only to the narrowest scope the new boundary needs.
