+++ scripts/gc_rekeyed_key_tables.json
@@ -98,7 +98,7 @@
       "why": "Cleared per dead object by gc_type_clear_dead_payload_side_tables (gc/types.rs:799) from the sweep (gc/oldgen.rs:700, 924, 1573), outside the dead_owner fan-out."
     },
     {
-      "site": "crates/perry-runtime/src/object/mod.rs::scan_transition_cache_slot",
+      "site": "crates/perry-runtime/src/object/side_table_roots.rs::scan_transition_cache_slot",
       "table": "TRANSITION_CACHE_GLOBAL",
       "death": "dead_owner:prune_dead_transition_cache_entries",
       "why": "#8192: registered prune drops the whole cache entry when either weak half (prev_keys keys-array, key_ptr interned string) is dead; next_keys is a strong root and cannot be. #9754: the per-slot body shared by the full walk and the young-log walk."
